+++ src/components/RestaurantDetailModal.tsx
@@ -1,4 +1,3 @@
-import { useEffect, useCallback } from "react";
 import CategoryIcon from "./CategoryIcon";
 import ModalPortal from "./common/ModalPortal";
 import type { Restaurant } from "../types/restaurant";
@@ -14,23 +13,6 @@ const RestaurantDetailModal = (props: RestaurantDetailModalProps) => {
   const { restaurant, onClose } = props;
   const { name, category, distance, description, link } = restaurant;
 
-  const onKeydown = useCallback(
-    (event: KeyboardEvent) => {
-      if (event.key === "Escape") {
-        onClose();
-      }
-    },
-    [onClose]
-  );
-
-  useEffect(() => {
-    window.addEventListener("keydown", onKeydown);
-
-    return () => {
-      window.removeEventListener("keydown", onKeydown);
-    };
-  }, [onKeydown]);
-
   return (
     <ModalPortal onClose={onClose}>
       <div className={styles.container}>
@@ -38,7 +20,7 @@ const RestaurantDetailModal = (props: RestaurantDetailModalProps) => {
         <div className={styles.info}>
           <h2 className={`${styles.name} text-title`}>{name}</h2>
           <span className={`${styles.distance} text-body`}>캠퍼스부터 {distance}분 내</span>
-          <p className={`${styles.description} text-body`}>{description ?? "현재 음식점에 대한 설명이 없습니다."}</p>
+          <p className={`${styles.description} text-body`}>{description}</p>
           <a href={link} className={styles.link} target="__blank">
             {link}
           </a>
